+++ internal/provider/provider.go
@@ -162,6 +162,7 @@ func (p *ONTAPProvider) Resources(ctx context.Context) []func() resource.Resourc
 		NewStorageVolumeResource,
 		NewStorageVolumeSnapshotResource,
 		NewSvmResource,
+		NewSVMPeersResource,
 	}
 }